# crawlers/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Linkedin parser API",
      "version": "0.0.1",
      "description": "Linkedin parser API"
    },
    "paths": {
...
      "/account": {
        "post": {
          "summary": "Add Linkedin account.",
          "parameters": [
            {
                "name": "Authorization",
                "in": "header",
                "description": "Auth token",
                "required": true,
                "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "description": "body",
              "type": "object",
              "required": true,
              "example": {
                  "input_data": {
                    "login": "linkedinlogin@gmail.com",
                    "li_at": "A12edwadcwe4rfwecfrtg45gre",
                    "password": "mypass1234"
                  }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      },
      "/accounts": {
        "get": {
          "summary": "Get all your Linkedin accounts.",
          "parameters": [
            {
                "name": "Authorization",
                "in": "header",
                "description": "Auth token",
                "required": true,
                "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      },
...
      "/search/scribe": {
        "post": {
          "summary": "Scribe from Linkedin search EACH profile.",
          "parameters": [
            {
                "name": "Authorization",
                "in": "header",
                "description": "Auth token",
                "required": true,
                "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "description": "body",
              "required": true,
              "type": "object",
              "example": {
                  "credentials_id": "123",
                  "input_data": {
                      "campaign_data": {
                          "search_url": "www.linkedin.com/user/search=marketers",
                          "interval_pages": 10
                      }
                  }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      }, 
...
      "/sn/scribe": {
        "post": {
          "summary": "Scribe Linkedin Sales Navigator member page.",
          "parameters": [
            {
                "name": "Authorization",
                "in": "header",
                "description": "Auth token",
                "required": true,
                "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "description": "body",
              "required": true,
              "type": "object",
              "example": {
                  "credentials_id": "123",
                  "input_data": {
                      "prospect_data": {
                          "linkedin": "www.linkedin.com/user/"
                      }
                  }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      }
    },
...
    "consumes": [
      "application/json"
    ]
  }  